swap content of home and profile screen
